+++ src/datetime/TimeRanges.cpp
@@ -23,7 +23,7 @@ namespace datetime {
 
 bool TimeRanges::IsActive() const
 {
-	for(int i=0;i<size(); i++)
+	for(size_t i=0;i<size(); i++)
 		if(at(i).IsActive())
 			return true;
 	return false;
